Don't fail files with pylint: skip-file
When pylint completely skips a file (e.g. with `# pylint: skip-file`),
there is no score output for that file because there is no output from
pylint at all.

The previous code considered this to be a failure because _parse_score
returns 0.0 if the score regex is not matched.

This code checks for cases where pylint produces no output and its
return code is 0, indicating that it intentionally skipped the
file. This commit introduces a new status of 'SKIPPED', which does not
cause the commit to be stopped.